Move data source manageable feature flag to DSM plugin
