+++ src/plugins/data_source/common/data_sources/types.ts
@@ -60,9 +60,3 @@ export enum DataSourceEngineType {
   Elasticsearch = 'Elasticsearch',
   NA = 'No Engine Type Available',
 }
-
-export enum ManageableBy {
-  All = 'all',
-  DashboardAdmin = 'dashboard_admin',
-  None = 'none',
-}
