+++ src/plugins/data_source_management/common/index.ts
@@ -6,3 +6,4 @@
 export const PLUGIN_ID = 'dataSourceManagement';
 export const PLUGIN_NAME = 'Data sources';
 export const DEFAULT_DATA_SOURCE_UI_SETTINGS_ID = 'defaultDataSource';
+export * from './types';
