+++ src/plugins/data_source_management/opensearch_dashboards.json
@@ -6,5 +6,6 @@
   "requiredPlugins": ["management", "indexPatternManagement"],
   "optionalPlugins": ["dataSource"],
   "requiredBundles": ["opensearchDashboardsReact", "dataSource", "opensearchDashboardsUtils"],
-  "extraPublicDirs": ["public/components/utils"]
+  "extraPublicDirs": ["public/components/utils"],
+  "configPath": ["data_source_management"]
 }
